requirements.txt
streamlit
langchain_groq
langchain
jsons
logging
pathlib
streamlit_mic_recorder
uuid
requests
bcrypt
python-dotenv
pymongo
